+++ src/modules/esl-utils/misc/test/uid.test.ts
@@ -1,4 +1,4 @@
-import {randUID, resetSequentialUID, sequentialUID} from '../uid';
+import {randUID, sequentialUID} from '../uid';
 
 const TRY_COUNT = 5;
 
@@ -20,18 +20,5 @@ describe('misc/uid helper tests', () => {
       set.add(sequentialUID('b'));
     }
     expect(set.size).toBe(TRY_COUNT * 2);
-  });
-
-  test('reset seqUID', () => {
-    const set = new Set();
-    for (let i = 0; i < TRY_COUNT; ++i) {
-      set.add(sequentialUID('c'));
-    }
-    expect(set.size).toBe(TRY_COUNT);
-    resetSequentialUID('c');
-    for (let i = 0; i < TRY_COUNT; ++i) {
-      set.add(sequentialUID('c'));
-    }
-    expect(set.size).toBe(TRY_COUNT);
   });
 });