+++ src/modules/esl-utils/misc/uid.ts
@@ -1,4 +1,12 @@
-const sequences = new Map<string, number>();
+const ns = window || global;
+const sequences = ns.__esl_sequences__ || new Map<string, number>();
+ns.__esl_sequences__ = sequences;
+
+declare global {
+  interface Window {
+    __esl_sequences__: Map<string, number>;
+  }
+}
 
 /** Create and return sequential id */
 export const sequentialUID = (name: string, prefix: string = name) => {
@@ -7,20 +15,11 @@ export const sequentialUID = (name: string, prefix: string = name) => {
   return prefix + uid;
 };
 
-/** Reset {@link sequentialUID} generator */
-export const resetSequentialUID = (name?: string) => {
-  if (typeof name === 'string') {
-    sequences.delete(name);
-  } else {
-    sequences.clear();
-  }
-};
-
 /** Return random unique identifier */
-export const randUID = (): string => {
+export const randUID = (prefix: string = ''): string => {
   const time = Date.now().toString(32);
   const rand = Math.round(Math.random() * 1024 * 1024).toString(32);
-  return time + '-' + rand;
+  return prefix + time + '-' + rand;
 };
 
 /**